Remove the call to ensureIOManagerIsRunning
It's not known why this line was added or if it works around a known issue, but System.Timeout.timeout does not call it, so we shouldn't need to either.  This line did not appear to reduce performance, but looking at the source code, it looks like it does a lot, and could lead to bad race conditions if called frequently.  Creating and using delays should be fast and safe, and should only poll for exceptions when necessary.